prims DIM O HECK YES BABYYYYYYYYY

+++ Node.java
@@ -1,7 +1,7 @@
 public class Node {
   public int index; 
-  public int weight; 
-  public Node (int index, int weight) {
+  public double weight; 
+  public Node (int index, double weight) {
     this.index = index;
     this.weight = weight;
   }
